191022|hashlib&many_to_many_model

+++ 03_Django/06_django_form/articles/views.py
@@ -2,14 +2,28 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article,Comment
 from .forms import ArticleForm, CommentForm
 from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
+import hashlib
 from IPython import embed
 # Create your views here.
 
 def index(request):
+    # 1. sessions 정보에서 visits_num 이라는 키로 접근해 값을 가져온다
+    # 해당하는 키가 없으면 0을 가져옴
+    # embed()
+    visits_num = request.session.get('visits_num',0)
+
+    # 2. 가져온 값을 sessions에 'visits_nmum'이라는 새로운 키의 값으로 1식 증가
+    request.session['visits_num'] = visits_num + 1
+
+    # 3. sessions data를 수정하면 Django는 수정한 내용을 알 수 없어서 작성하는 구문
+    request.session.modified = True
+    
     articles = Article.objects.all()
-    context = {'articles':articles,}
+    context = {'articles':articles,'visits_num': visits_num, }
     return render(request, 'articles/index.html', context)
 
+@login_required
 def create(request):
     """
     Form Class
@@ -26,7 +40,9 @@ def create(request):
         # 해당 폼이 유효한지 확인
         if form.is_valid():
             # form.cleaned_data를 통해 폼 데이터를 정제한다.(type(form.cleaned_data) = dict)
-            article = form.save()
+            article = form.save(commit=False)
+            article.user_id = request.user.id
+            article.save()
 
             return redirect('articles:detail', article.pk )
     else:
@@ -45,21 +61,27 @@ def detail(request,article_pk):
 
 @require_POST
 def delete(request,article_pk):
-    article = Article.objects.get(pk=article_pk)
-    article.delete()
+    if request.user.is_authenticated:
+        article = Article.objects.get(pk=article_pk)
+        if request.user == article.user:
+            article.delete()
     return redirect('articles:index')
 
 
+@login_required
 def update(request,article_pk):
     article = get_object_or_404(Article,pk=article_pk)
-    if request.method == 'POST':
-        # instance -> 수정의 대상이 되는 특정한 글 객체
-        form = ArticleForm(request.POST,instance=article)
-        if form.is_valid():
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.user == article.user:
+        if request.method == 'POST' :
+            # instance -> 수정의 대상이 되는 특정한 글 객체
+            form = ArticleForm(request.POST,instance=article)
+            if form.is_valid():
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:detail', article.pk)
 
     context = {'form':form,'article':article,}
     return render(request, 'articles/form.html', context)
@@ -77,19 +99,23 @@ Update 로직
 2.POST
  - 수정된 글을 DB에 저장
 """
+
 @require_POST
 def comments_create(request,article_pk):
     # article = get_object_or_404(Article, pk=article_pk)
-    commentform = CommentForm(request.POST)
-    if commentform.is_valid():
-        comment = commentform.save(commit=False)
-        comment.article_id = article_pk
-        comment.save()
+    if request.user.is_authenticated:
+        commentform = CommentForm(request.POST)
+        if commentform.is_valid():
+            comment = commentform.save(commit=False)
+            comment.article_id = article_pk
+            comment.user_id = request.user.id
+            comment.save()
     return redirect('articles:detail', article_pk)
-
 
 @require_POST
 def comments_delete(request,article_pk,comment_pk):
-    comment = get_object_or_404(Comment, pk=comment_pk)
-    comment.delete()
+    if request.user.is_authenticated:
+        comment = get_object_or_404(Comment, pk=comment_pk)
+        if comment.user == request.user:
+            comment.delete()
     return redirect('articles:detail', article_pk)
